Use bindable properties and refactor views
