+++ app/src/main/java/com/skydoves/themovies2/view/ui/details/tv/TvDetailActivity.kt
@@ -39,7 +39,6 @@ class TvDetailActivity :
     super.onCreate(savedInstanceState)
     binding {
       activity = this@TvDetailActivity
-      lifecycleOwner = this@TvDetailActivity
       viewModel = vm.apply { postTvId(intentTv.id) }
       tv = intentTv
       videoAdapter = VideoListAdapter()
